updated jdg-remote quickstarts to align with deploying to wildfly and project scripts
