+++ jdg-remote-cache-materialization/kits/wildfly/modules/com/client/quickstart/addressbook/remotecache/pojos/main/module.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module xmlns="urn:jboss:module:1.0" name="com.client.quickstart.addressbook.remotecache.pojos">
+    <resources>
+        <resource-root path="jdg-remote-cache-materialization-pojos.jar" />
+        <!-- Insert resources here -->
+    </resources>
+
+    <dependencies>
+        <module name="org.infinispan.client.hotrod" slot="${jdg.slot}" optional="true" services="export"/>
+   	<module name="org.infinispan.protostream" slot="${jdg.slot}"  optional="true" services="export"/>
+
+        <module name="org.jboss.teiid.resource-adapter.infinispan.hotrod" export="true" />
+    </dependencies>
+</module>
