+++ jdg-remote-cache-materialization/src/main/java/com/client/quickstart/addressbook/remotecache/pojos/marshallers/PhoneNumberMarshaller.java
@@ -1,0 +1,56 @@
+/*
+ * JBoss, Home of Professional Open Source
+ * Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
+ * contributors by the @authors tag. See the copyright.txt in the
+ * distribution for a full listing of individual contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.client.quickstart.addressbook.remotecache.pojos.marshallers;
+
+import org.infinispan.protostream.MessageMarshaller;
+import com.client.quickstart.addressbook.remotecache.pojos.domain.PhoneNumber;
+import com.client.quickstart.addressbook.remotecache.pojos.domain.PhoneType;
+
+import java.io.IOException;
+
+/**
+ * @author Adrian Nistor
+ */
+public class PhoneNumberMarshaller implements MessageMarshaller<PhoneNumber> {
+
+   @Override
+   public String getTypeName() {
+      return "quickstart.Person.PhoneNumber";
+   }
+
+   @Override
+   public Class<PhoneNumber> getJavaClass() {
+      return PhoneNumber.class;
+   }
+
+   @Override
+   public PhoneNumber readFrom(ProtoStreamReader reader) throws IOException {
+      String number = reader.readString("number");
+      PhoneType type = reader.readObject("type", PhoneType.class);
+
+      PhoneNumber phoneNumber = new PhoneNumber();
+      phoneNumber.setNumber(number);
+      phoneNumber.setType(type);
+      return phoneNumber;
+   }
+
+   @Override
+   public void writeTo(ProtoStreamWriter writer, PhoneNumber phoneNumber) throws IOException {
+      writer.writeString("number", phoneNumber.getNumber());
+      writer.writeObject("type", phoneNumber.getType(), PhoneType.class);
+   }
+}
